src/services/graph_service.rs
use crate::AppState;
use crate::models::graph::{GraphData, Edge};

pub struct GraphService;

impl GraphService {
    pub fn get_graph_data(state: &AppState) -> Result<GraphData, std::io::Error> {
        // Implementation goes here
        Ok(GraphData::default())
    }

    pub fn refresh_graph_data(state: &AppState) -> Result<GraphData, std::io::Error> {
        // Implementation goes here
        Ok(GraphData::default())
    }

    pub fn build_edges(state: &AppState) -> Result<Vec<Edge>, std::io::Error> {
        // Implementation goes here
        Ok(Vec::new())
    }
}
